refactor: query에서 command로 일부 메서드 옮김
