+++ src/main/java/coffeemeet/server/certification/service/cq/CertificationCommand.java
@@ -5,6 +5,7 @@ import coffeemeet.server.certification.domain.CompanyEmail;
 import coffeemeet.server.certification.domain.Department;
 import coffeemeet.server.certification.repository.CertificationRepository;
 import coffeemeet.server.user.domain.User;
+import java.util.function.Consumer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,6 +14,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 @RequiredArgsConstructor
 public class CertificationCommand {
+
+  private static final String EXISTED_COMPANY_EMAIL = "이미 사용 중인 회사 이메일입니다.";
 
   private final CertificationRepository certificationRepository;
 
@@ -26,6 +29,16 @@ public class CertificationCommand {
             .user(user)
             .build()
     );
+  }
+
+  public void hasDuplicatedCompanyEmail(CompanyEmail companyEmail) {
+    if (certificationRepository.existsByCompanyEmail(companyEmail)) {
+      throw new IllegalArgumentException(EXISTED_COMPANY_EMAIL);
+    }
+  }
+
+  public void applyIfCertifiedUser(Long userId, Consumer<? super Certification> consumer) {
+    certificationRepository.findByUserId(userId).ifPresent(consumer);
   }
 
 }
